Add opt-out for UA parsing
Currently the only option for caching responses from CSP are to use
`setAllHeaders: true`, lest a browser-specific CSP gets cached for the wrong
UA. This PR adds a new option, `setLegacyHeaders`, which will skip UA parsing
if defined and falsey.

This adds a testing dependency on `lodash.pickBy`, which we can remove by
upgrading lodash to the latest version.

Lastly, I added a constants file for header names in `lib/headers.js`. In
order to minimize change, I didn't refactor existing files to use it, but it
might help cut down on duplication and human error potential to use it in
other places.

Fixes #32.